# src/main/java/ru/sarvartdinov/sber/api/GroupsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ru.sarvartdinov.sber.api;

import ru.sarvartdinov.sber.api.model.Group;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-01-21T12:53:25.046+05:00")

@Api(value = "groups", description = "the groups API")
public interface GroupsApi {

    GroupsApiDelegate getDelegate();

    @ApiOperation(value = "Get groups by ids", nickname = "getGroups", notes = "Ендпоинт для получения информации по указанным группам", response = Group.class, responseContainer = "List", tags={ "groupApi", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Найденные группы", response = Group.class, responseContainer = "List"),
        @ApiResponse(code = 404, message = "Группы с запрошенными идентификаторами не найдены") })
    @RequestMapping(value = "/groups",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Group>> getGroups(@NotNull @ApiParam(value = "Список идентификаторов групп, по которым необходимо отдать информацию", required = true) @Valid @RequestParam(value = "groupIds", required = true) List<String> groupIds) {
        return getDelegate().getGroups(groupIds);
    }

}